+++ soroban-sdk/src/lib.rs
@@ -65,7 +65,7 @@ mod account;
 mod bigint;
 mod bytes;
 mod contract_data;
-mod events;
+mod contract_event;
 pub mod iter;
 mod ledger;
 mod map;
@@ -77,7 +77,7 @@ pub use bigint::BigInt;
 pub use bytes::{Binary, FixedBinary};
 pub use bytes::{Bytes, BytesN};
 pub use contract_data::ContractData;
-pub use events::Events;
+pub use contract_event::Events;
 pub use ledger::Ledger;
 pub use map::Map;
 pub use set::Set;
